Update lab 7
